requirements/requirements-ci-freeze.txt
alabaster==0.7.13
astroid==2.15.8
Babel==2.12.1
bcrypt==4.0.1
build==1.0.3
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
check-manifest==0.49
colorama==0.4.6
coverage==7.3.1
cryptography==41.0.4
devpi-client==6.0.5
devpi-common==3.7.0
dill==0.3.7
distlib==0.3.7
docutils==0.20.1
exceptiongroup==1.1.3
filelock==3.12.4
flake8==6.1.0
idna==3.4
imagesize==1.4.1
importlib-metadata==6.8.0
iniconfig==2.0.0
isort==5.12.0
jaraco.classes==3.3.0
Jinja2==3.1.2
keyring==24.2.0
lazy==1.6
lazy-object-proxy==1.9.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.1.0
mypy==1.5.1
mypy-extensions==1.0.0
nh3==0.2.14
packaging==23.2
paramiko==3.3.1
pep517==0.13.0
pkginfo==1.9.6
platformdirs==3.10.0
pluggy==1.3.0
py==1.11.0
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
Pygments==2.16.1
pylint==2.17.7
PyNaCl==1.5.0
pyproject-api==1.6.1
pyproject_hooks==1.0.0
pytest==7.4.2
pytest-runner==6.0.0
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
readme-renderer==42.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.6.0
ruamel.yaml==0.17.33
ruamel.yaml.clib==0.2.7
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
tomli==2.0.1
tomlkit==0.12.1
tox==4.11.3
twine==4.0.2
types-pytz==2023.3.1.1
types-tzlocal==4.3.0.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.5
virtualenv==20.24.5
wrapt==1.15.0
zipp==3.17.0
